Fixed some problems and added a .obj loader to the OpenGL widget
There was a problem where the interface crashed due to a concurrent replot and modification of the histograms data due to these methods not being thread safe. It should have been working but I put a not where it should have been a ! to check if the previous replot was finished, thanks python.
Setup the QOpenGLWidget to render simple .obj files that can be loaded from disk. These models can include a texture. The OpenGL view can be rotated, panned and zoomed in and out.