Coded the OpenGL 3D view to show the DCS setup
The QOpenGL widget now shows both the parallel and antiparallel configurations (stacked vertically). We can zoom, orbit and pan the view. All 3D models and textures are stored in separate files so we can modify it after compiling. All angles change according to the input parameters and the second crystal angle changes during the simulation to represent the scanning angle.