Added the folder with the input files for the simulation
Added the folder where we can change various input parameters, such as the reflection profiles. This folder also has the 3D models and textures for the OpenGL view.
An extra file where we can specify the location of the folder with all these files in your system. This file has to be next to the compiled executable to be read, this should be the only hardcoded path in the program.