Fixed a problem outputting data to files
Found a problem where the output files were not being opened and we could not get the simulated data.
Also found a problem where the image plate for the detector in the antiparallel configuration was not being written. The data for this plate was being added to the second crystal in antiparallel by mistake.