Added the use of multiple crystal response functions for each energy
Modified the reading of the crystal response to read a set of responses for various energies.
The reflectivity now is performed taking into account each ray's energy, selecting the appropriate response function and (linearly) interpolating the reflectivity of the response for the energy right below and right above the incident.