Restructure and CUDA support
Restructured the directories for the project.
Finally managed to compile the code with the CUDA language.
Ended up making a separate static library for the device code that is linked with the main simulation (host) code.
MSVC cmake is crap. The CUDA support does not work properly.
After 2 days I finally managed to find the variable that caused the NVCC -dlink command to be malformed.
The -ccbin flag that points to the host CXX compiler was not expanding properly (only in the dlink command), leaving the token <CMAKE_CUDA_HOST_COMPILER> in the command.
The flag was meant for extra link options, which in MSVC was:
_CMAKE_CUDA_EXTRA_FLAGS = -forward-unknown-to-host-compiler -ccbin=<CMAKE_CUDA_HOST_COMPILER>
I had to set using:
set(_CMAKE_CUDA_EXTRA_FLAGS "-forward-unknown-to-host-compiler -ccbin=\"${CMAKE_CXX_COMPILER}\"")