GPU Accelerated Simulation
Added first support for the GPU simulation.
Each bin is parallelized on the GPU.
The management between bins are still done in the CPU, using the same code as for OpenMP and no parallelization.
No profiling and performance tests have been done.
This will be done in the following commits.