Fixed Reflectivity profiles calculation
Tweaked a few values on output files and added some more information.
Fixed a problem with the reflection profiles, where the total reflection probability was higher then 1.
This resulted in a probability distribution that was cut at the peak. It was also preventing us from compraring different profiles.
Now all profiles are normalized by the max reflectivity for every energy and this constant is reported in the general output.
The counts for each histogram are also multiplied by this constant, giving us a more absolute value of the counts.
This also enables us to compare simulations with different profiles, i.e. different crystals, different f1 or f2 values, etc...