Fixed the events visualizer
The order of rotations was switched when reseting the event line referencial in the qdcs3dvis.cpp file.
Also fixed a strange bug in the gif output where the image vector variable was not keeping its size when cleared, resulting in an index out of bounds, but only in debug mode.